requirements.txt
streamlit
pandas
numpy
Pillow
plotly
scikit-learn
tensorflow
fpdf